SOLR-7954: Fixed an integer overflow bug in the HyperLogLog code used by the 'cardinality' option of stats.field to prevent ArrayIndexOutOfBoundsException in a distributed search when a large precision is selected and a large number of values exist in each shard (merge r1697969)
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/branch_5x@1697977 13f79535-47bb-0310-9956-ffa450edef68

+++ solr/core/src/java/org/apache/solr/util/hll/BigEndianAscendingWordSerializer.java
@@ -85,7 +85,7 @@ class BigEndianAscendingWordSerializer implements IWordSerializer {
         this.wordLength = wordLength;
         this.wordCount = wordCount;
 
-        final long bitsRequired = (wordLength * wordCount);
+        final long bitsRequired = ((long)wordLength) * ((long)wordCount);
         final boolean leftoverBits = ((bitsRequired % BITS_PER_BYTE) != 0);
         final int bytesRequired = (int)(bitsRequired / BITS_PER_BYTE) + (leftoverBits ? 1 : 0) + bytePadding;
         bytes = new byte[bytesRequired];
